python/azmlv1_dataset_operation/pyproject.toml
[tool.poetry]
name = "azmlv1-dataset-operation"
version = "0.1.0"
description = ""
authors = ["Taiki Sugawara <buzz.taiki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
azureml-core = "^1.38.0"
azureml-dataset-runtime = {version = "^1.43.0", platform = "linux"}
azure-cli-core = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
